Main changes: - Added logic to check product availability (availAmount) when adding to the cart. - Implemented automatic deduction of product balances after successful payment (Stripe). - In `ProductService` and `PaymentService`, identifier types are converted to Long.
Bug fixes:
- Fixed `LazyInitializationException` for product images in the cart.
- Fixed `AuthenticationException` error with Stripe API key.
- Added correct price display with discounts in the cart (`cart.html`).

UI/UX:
- Fixed footer positioning (Sticky Footer) on the Home, Cart, and Product View pages.
- Fixed a visual bug with the background bar in the footer (z-index fix).

+++ src/main/java/com/example/techzone/controller/web/FeedbackController.java
@@ -23,7 +23,7 @@ public class FeedbackController {
 
     @PostMapping("/review")
     public String addReview(@ModelAttribute("review") @Valid FeedbackCreateDto dto,
-                            @PathVariable int productId,
+                            @PathVariable Long productId,
                             Principal principal) {
         if (principal == null) {
             return "redirect:/auth/login";
